requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
dataclasses-json==0.6.7
docx2txt==0.8
faiss-cpu==1.8.0.post1
frozenlist==1.4.1
gigachain==0.2.16
gigachain-core==0.2.38.post2
gigachain-text-splitters==0.2.4
gigachat==0.1.35
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.8
langchain-text-splitters==0.3.0
langfuse==2.51.2
langsmith==0.1.130
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-telegram-bot==21.6
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tenacity==8.5.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.13.1